% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln_to_base_and_position_table.R
\name{aln_to_base_and_position_table}
\alias{aln_to_base_and_position_table}
\title{Alignment to Base and Position Table}
\usage{
aln_to_base_and_position_table(aln)
}
\arguments{
\item{aln}{A single row data frame.}
}
\value{
A data frame with base, position, and qname.
}
\description{
Takes a single row produced by filter_bam_alignments and returns a table with base, position, and qname.
}
\details{
input is a single row produced by filter_bam_alignments
otherwise input must have seq, pos, qwidth, and qname columns
}
