% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_qualitative_p_value_score.R
\name{assign_qualitative_p_value_score}
\alias{assign_qualitative_p_value_score}
\title{Assign Qualitative P-Value Score}
\usage{
assign_qualitative_p_value_score(x)
}
\arguments{
\item{x}{A numeric value representing the p-value.}
}
\value{
A character string representing the significance level:
\itemize{
\item \verb{***} for p-values less than 0.001
\item \verb{**} for p-values less than 0.01
\item \code{*} for p-values less than 0.05
\item An empty string (\code{''}) for p-values greater than or equal to 0.05
}
}
\description{
This function assigns a qualitative score based on a given p-value.
It uses asterisks to represent the significance level, following common conventions in statistical reporting.
}
