% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_file_to_table.R
\name{bam_file_to_table}
\alias{bam_file_to_table}
\title{BAM File to Table}
\usage{
bam_file_to_table(
  bam_file,
  scan_param_what_list = c("qname", "rname", "strand", "pos", "qwidth", "seq", "cigar",
    "flag")
)
}
\arguments{
\item{bam_file}{A character string specifying the path to the BAM file.}

\item{scan_param_what_list}{A character vector specifying the scan parameters.}
}
\value{
A data frame containing information from the BAM file.
}
\description{
Take in a BAM file, read in with scanBam, and tag each unique BAM file with the filepath's basename.
}
