% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_complement.R
\name{base_complement}
\alias{base_complement}
\title{Base Complement}
\usage{
base_complement(x)
}
\arguments{
\item{x}{A single character string representing a DNA base ('A', 'T', 'G', 'C', or 'N').}
}
\value{
A single character string representing the complementary base or N if input is N.
}
\description{
This function takes a single DNA base and returns its complementary base according to standard base pairing rules.
}
