% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_base_percentages.R
\name{calculate_base_percentages}
\alias{calculate_base_percentages}
\title{Calculate Base Percentages}
\usage{
calculate_base_percentages(
  df_aln_pos,
  edit_site_position,
  edit_site_strand,
  bases_from_cut_site = 20
)
}
\arguments{
\item{df_aln_pos}{A data frame with base, position, and qname.}

\item{edit_site_position}{The position of the edit site.}

\item{edit_site_strand}{The strand of the edit site ('+' or '-').}

\item{bases_from_cut_site}{Number of bases to consider from the cut site.}
}
\value{
A data frame with counts and percentages of each base at each position. The number of aligned reads (depth) for each position is also reported.
}
\description{
Uses input from \code{parallel_aln_to_base_and_position_tables()} or \code{aln_to_base_and_position_tables()}  and returns a table with counts and percentages of each base at each position.
}
