% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reference_genome.R
\name{check_reference_genome}
\alias{check_reference_genome}
\title{Check Reference Genome Path}
\usage{
check_reference_genome(run_params)
}
\arguments{
\item{run_params}{A list or data frame containing the run parameters. It
should include a field \code{reference_genome_path} which is the path to the
reference genome file.}
}
\value{
NULL. This function is used for its side effect of stopping the
execution if the reference genome path does not exist.
}
\description{
This function checks whether the reference genome path specified in the
run parameters exists. If the reference genome path is not NULL and does
not exist, the function will stop and return an error message.
}
\examples{
\dontrun{
# Example run parameters
run_params <- list(reference_genome_path = "path/to/reference_genome.fasta")

# Check if the reference genome path exists
check_reference_genome(run_params)
}

}
