% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_reference_bases.R
\name{extract_reference_bases}
\alias{extract_reference_bases}
\title{Extract Reference Bases for Edit Sites}
\usage{
extract_reference_bases(genomic_sequence, df_base_percentages)
}
\arguments{
\item{genomic_sequence}{A \code{BSgenome} object containing the genomic sequence from which reference bases will be extracted.}

\item{df_base_percentages}{A data frame containing base percentages and related information for edit sites.}
}
\value{
A data frame with reference bases integrated into the input \code{df_base_percentages}.
}
\description{
This function extracts reference bases for specified edit sites from a genomic sequence and integrates them into a data frame containing base percentages.
}
