% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bam_files.R
\name{list_bam_files}
\alias{list_bam_files}
\title{List BAM Files}
\usage{
list_bam_files(base_directory)
}
\arguments{
\item{base_directory}{A string specifying the base directory that points to the iGUIDE file structure containing BAM alignments.}
}
\value{
A 1-D matrix of BAM file paths.
}
\description{
This function lists BAM alignment files from the specified iGUIDE file structure.
It ignores any files ending with \code{.unsorted.bam}.
}
