% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference_genome.R
\name{load_reference_genome}
\alias{load_reference_genome}
\title{Load Reference Genome}
\usage{
load_reference_genome(reference_genome_path)
}
\arguments{
\item{reference_genome_path}{A string specifying the path to the reference genome file.}
}
\value{
A \code{DNAStringSet} object containing the reference genome sequences.
}
\description{
This function loads a reference genome from a specified file path.
}
\examples{
\dontrun{
reference_genome_path <- "path/to/reference_genome.fasta"
reference_genome <- load_reference_genome(reference_genome_path)
}
}
