% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edit_site_table.R
\name{make_edit_site_table}
\alias{make_edit_site_table}
\title{Make Edit Sites Table}
\usage{
make_edit_site_table(
  ft_data_table,
  spec_info_combo_overview_table,
  abundance_cutoff = 3
)
}
\arguments{
\item{ft_data_table}{A data table containing information about edit sites.}

\item{spec_info_combo_overview_table}{A data table containing information about specimen and annotations.}

\item{abundance_cutoff}{Numeric value specifying the abundance cutoff for filtering.}
}
\value{
A data frame containing information about edit sites.
}
\description{
Description: Makes Edit sites dataframe. This is the experimental object that is used to identify
iGUIDE edit sites in alignment data
}
