% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_aln_to_base_and_position_tables.R
\name{parallel_aln_to_base_and_position_tables}
\alias{parallel_aln_to_base_and_position_tables}
\title{Parallel Alignment to Base and Position Tables}
\usage{
parallel_aln_to_base_and_position_tables(df_bam_filtered, number_of_cpu = 4)
}
\arguments{
\item{df_bam_filtered}{A data frame of filtered BAM alignments.}

\item{number_of_cpu}{Number of CPU cores to use for parallel processing.}
}
\value{
A data frame with base, position, and qname.
}
\description{
Runs aln_to_base_and_position_table on each row of a data frame produced by filter_bam_alignments.
Produces a large long-form data frame of base, position, and qname.
}
