% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_relative_to_cut_site.R
\name{position_relative_to_cut_site}
\alias{position_relative_to_cut_site}
\title{Compute Position Relative to Cut Site}
\usage{
position_relative_to_cut_site(df_edits)
}
\arguments{
\item{df_edits}{A data frame containing the edit site information.
It should include columns such as \code{specimen}, \code{edit_site_target.seq}, \code{edit.site},
\code{edit_site_strand}, and \code{position}.}
}
\value{
A data frame with an additional column, \code{relative_position}, representing the
position of each edit relative to the maximum position within each group of
\code{specimen}, \code{edit_site_target.seq}, and \code{edit.site}. If the strand is '+',
the relative position is calculated as \code{position - max(position)}, otherwise it is
calculated as \code{max(position) - position}.
}
\description{
This function calculates the position of bases relative to a cut site (position 0) in an iGUIDE dataset.
If a cut is on the plus strand, then it is numbered negatively to 0.
}
