% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_combo_overview_table.R
\name{pull_combo_overview_table}
\alias{pull_combo_overview_table}
\title{Pull Combo Overview Table}
\usage{
pull_combo_overview_table(base_directory)
}
\arguments{
\item{base_directory}{A string specifying the base directory that points to the iGUIDE file structure.}
}
\value{
A dataframe containing the \code{combo_overview} table.
}
\description{
This function retrieves the \code{combo_overview} table from the \code{spec_info} section
of an iGUIDE file structure.
}
\details{
The function expects a base directory that points to the iGUIDE file structure.
The iGUIDE output structure must be present for the function to work correctly.
The output is a dataframe that links GTSPs to annotations.
}
