% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_ft_data_tables.R
\name{pull_ft_data_tables}
\alias{pull_ft_data_tables}
\title{Pull FT Data Tables}
\usage{
pull_ft_data_tables(base_directory)
}
\arguments{
\item{base_directory}{A string specifying the base directory that points to the iGUIDE file structure with \code{ft_data}.}
}
\value{
A dataframe containing the combined \code{ft_data} tables.
}
\description{
This function retrieves all \code{ft_data} tables found in a specified iGUIDE file structure.
}
