% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_edits.R
\name{quantify_edits}
\alias{quantify_edits}
\title{Quantify Edits in Sequencing Data}
\usage{
quantify_edits(
  base_directory,
  output_directory,
  analysis_name = "quantify_edits",
  abundance_cutoff = 3,
  end_distance_from_cut_site = 20,
  cut_site_start_distance_within_gRNA = 3,
  cut_site_start_distance_outside_gRNA = 3,
  reference_genome_path = "",
  editable_base = "A",
  expected_edit = "G",
  binomial_p_value_threshold = 0.05,
  binomial_direction = "greater",
  n_processors = 4,
  overwrite = TRUE
)
}
\arguments{
\item{base_directory}{Character. The base directory where input data is located.}

\item{output_directory}{Character. The directory where output data will be saved.}

\item{analysis_name}{Character. The name of the analysis (default: 'quantify_edits').}

\item{abundance_cutoff}{Numeric. The minimum abundance cutoff for including an edit site (default: 3).}

\item{end_distance_from_cut_site}{Numeric. The maximum distance from the cut site to consider bases (default: 20).}

\item{cut_site_start_distance_within_gRNA}{Numeric. The distance within gRNA to start considering cut sites (default: 3).}

\item{cut_site_start_distance_outside_gRNA}{Numeric. The distance outside gRNA to start considering cut sites (default: 3).}

\item{reference_genome_path}{Character. Path to a fasta file used to generate the base iGUIDE result (default: '').}

\item{editable_base}{Character. The base that is being edited (default: 'A').}

\item{expected_edit}{Character. The expected base after editing (default: 'G').}

\item{binomial_p_value_threshold}{Numeric. The p-value threshold for determining significance using a binomial test (default: 0.05).}

\item{binomial_direction}{Character. The direction of the binomial test ('greater', 'less', or 'two-sided') (default: 'greater').}

\item{n_processors}{Numeric. The number of processors to use for parallel processing (default: 4).}

\item{overwrite}{Logical. Whether to overwrite existing analysis output (default: TRUE).}
}
\value{
Writes a CSV and RDS file containing base percentages and statistical results to the specified output directory.
}
\description{
This function quantifies edits in sequencing data by processing BAM files, filtering alignments, and calculating base percentages at specific edit sites. It performs statistical testing to determine the significance of edits.
}
\examples{
\dontrun{
quantify_edits(
  base_directory = "path/to/base_directory",
  output_directory = "path/to/output_directory",
  analysis_name = "my_analysis",
  abundance_cutoff = 5,
  end_distance_from_cut_site = 25,
  cut_site_start_distance_within_gRNA = 4,
  cut_site_start_distance_outside_gRNA = 4,
  reference_genome_path = 'path/to/genome.fasta',
  editable_base = 'C',
  expected_edit = 'T',
  binomial_p_value_threshold = 0.01,
  binomial_direction = 'two-sided',
  n_processors = 6,
  overwrite = FALSE
)
}
}
