% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_edit_site_id.R
\name{split_edit_site_id}
\alias{split_edit_site_id}
\title{Split Edit Site ID into Chromosome, Strand, and Position}
\usage{
split_edit_site_id(df)
}
\arguments{
\item{df}{A data frame containing an \code{edit.site} column with coordinates in the format "chr:str:pos".}
}
\value{
A data frame with the original columns and three new columns: \code{chromosome}, \code{strand}, and \code{position}.
}
\description{
This function takes a data frame with an \code{edit.site} column containing coordinates
in the format "chr:str:pos" and splits it into separate columns for chromosome, strand, and position.
}
